# backend/src/swagger/swagger.yaml
openapi: 3.0.0
info:
  title: Pick-N-Get Rider API
  description: API for managing riders, pickups, and locations in the Pick-N-Get recycling platform
  version: 1.0.0
  contact:
    name: Pick-N-Get Team
    email: support@picknget.com
servers:
  - url: https://pick-n-get-be.onrender.com
    description: Production server
  - url: http://localhost:5000
    description: Development server
...
components:
  schemas:
...
    Product:
      type: object
      properties:
        productId:
          type: number
          description: Unique product identifier
        walletAddress:
          type: string
          description: Vendor's wallet address
        name:
          type: string
          description: Product name
        description:
          type: string
          description: Product description
        category:
          type: string
          enum:
            [
              Bags & Accessories,
              Furniture,
              Office Supplies,
              Fitness & Wellness,
              Home & Garden,
              Textiles,
              Electronics,
              Others,
            ]
        price:
          type: number
          description: Price in HBAR
        priceUSD:
          type: number
          description: Price in USD (for display)
        quantity:
          type: number
          description: Available quantity
        weight:
          type: number
          description: Weight in kg
        imageFileId:
          type: string
          description: Hedera File Service ID
        imageUrl:
          type: string
          description: Full image URL
        txHash:
          type: string
          description: Blockchain transaction hash
        status:
          type: string
          enum: [Available, Sold Out, Inactive]
        views:
          type: number
          description: Number of views
        sales:
          type: number
          description: Number of sales
        revenue:
          type: number
          description: Total revenue
        recycledPercentage:
          type: number
          description: Percentage of recycled material
        carbonNeutral:
          type: boolean
          description: Carbon neutral status
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
paths:
...
  /api/v1/products:
    post:
      tags:
        - Products
      summary: Add a new product
      description: Add product to marketplace (blockchain + backend)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - productId
                - walletAddress
                - name
                - description
                - category
                - price
                - quantity
                - weight
                - imageFileId
                - txHash
              properties:
                productId:
                  type: number
                  example: 1
                walletAddress:
                  type: string
                  example: "0x0e54c9054f000849a5f82d53914daaf97f306e95"
                name:
                  type: string
                  example: "Recycled Tote Bag"
                description:
                  type: string
                  example: "Eco-friendly bag made from recycled materials"
                category:
                  type: string
                  example: "Bags & Accessories"
                price:
                  type: number
                  example: 10
                  description: Price in HBAR
                priceUSD:
                  type: number
                  example: 0.5
                  description: Price in USD (optional)
                quantity:
                  type: number
                  example: 20
                weight:
                  type: number
                  example: 0.5
                  description: Weight in kg
                imageFileId:
                  type: string
                  example: "0.0.7179369"
                  description: Hedera File ID
                txHash:
                  type: string
                  example: "0x306c58ea..."
                  description: Blockchain transaction hash
                recycledPercentage:
                  type: number
                  example: 85
                carbonNeutral:
                  type: boolean
                  example: true
      responses:
        "201":
          description: Product added successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "success"
                  message:
                    type: string
                  data:
                    type: object
                    properties:
                      productId:
                        type: number
                      name:
                        type: string
                      price:
                        type: number
                      priceUSD:
                        type: number
                      imageUrl:
                        type: string
        "400":
          description: Missing required fields
        "403":
          description: Producer not registered
        "409":
          description: Product ID already exists
        "500":
          description: Internal server error

    get:
      tags:
        - Products
      summary: Get all products
      description: Retrieve products for shop page with filters
      parameters:
        - in: query
          name: status
          schema:
            type: string
            enum: [Available, Sold Out, Inactive]
            default: Available
        - in: query
          name: category
          schema:
            type: string
        - in: query
          name: limit
          schema:
            type: number
            default: 50
        - in: query
          name: offset
          schema:
            type: number
            default: 0
      responses:
        "200":
          description: Products retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  data:
                    type: object
                    properties:
                      products:
                        type: array
                        items:
                          $ref: "#/components/schemas/Product"
                      total:
                        type: number
                      limit:
                        type: number
                      offset:
                        type: number
...